% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_template.R
\name{get_template_scenarios}
\alias{get_template_scenarios}
\title{Get template scenarios}
\usage{
get_template_scenarios()
}
\value{
Vector of strings
}
\description{
get_template_scenarios returns the template scenario names.
}
