% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_template.R
\name{get_template_variables}
\alias{get_template_variables}
\title{Get template variable}
\usage{
get_template_variables(return_only_names = FALSE)
}
\arguments{
\item{return_only_names}{If TRUE (default set to FALSE), return vector of variable names.}
}
\value{
Data-frame with the reporting template variables
}
\description{
get_template_variables returns the template variable names and definitions.
}
