% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmmi-package.R
\docType{package}
\name{gmmi-package}
\alias{gmmi}
\alias{gmmi-package}
\title{gmmi: GMMI Tools: Data Handling for the GMMI Project}
\description{
Data Handling for the GMMI Project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/johanneskoch94/gmmi}
  \item Report bugs at \url{https://github.com/johanneskoch94/gmmi/issues}
}

}
\author{
\strong{Maintainer}: Johannes Koch \email{jokoch@pik-potsdam.de}

}
\keyword{internal}
