% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_over_time}
\alias{plot_over_time}
\alias{plot_diff_over_emi}
\alias{plot_carbon_prive_over_emi}
\title{Plot GMMI variables}
\usage{
plot_over_time(data, variable)

plot_diff_over_emi(data, variable = "GDP")

plot_carbon_prive_over_emi(data, years = 2025:2050)
}
\arguments{
\item{data}{data-frame with data}

\item{variable}{string with the variable name to plot}

\item{years}{years with which to filter the data before plotting}
}
\value{
ggplot
}
\description{
plot_over_time creates a plot of a GMMI variable over time.
}
\section{Functions}{
\itemize{
\item \code{plot_diff_over_emi()}: plot relative difference between scenarios over diff in emissions

\item \code{plot_carbon_prive_over_emi()}: plot carbon price over diff in emissions

}}
