% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_template.R
\name{read_template}
\alias{read_template}
\title{Read-in and tidy the reporting template.}
\usage{
read_template()
}
\value{
Tidy data-frame with the reporting template
}
\description{
read_template reads-in the template file (shipped with the package), and performs minor operations to tidy up
the data.
}
