% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_submission.R
\name{read_submission}
\alias{read_submission}
\title{Read in data from GMMI submissions}
\usage{
read_submission(file, verbose = FALSE, save_log = FALSE)
}
\arguments{
\item{file}{Path or vector of file paths.}

\item{verbose}{If TRUE (FALSE by default), print information to screen.}

\item{save_log}{If TRUE (FALSE by default), redirect output to "read_submission.log"}
}
\value{
data-frame with submission data
}
\description{
Read in data from one or multiple GMMI submissions.
}
